feat: 🎸 Adjust PointController API's and create unit tests

+++ Vale.Geographic/Vale.Geographic.Application.Core/Services/PointOfInterestAppService.cs
@@ -1,43 +1,47 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using AutoMapper;
-using NetTopologySuite;
-using NetTopologySuite.Geometries;
-using NetTopologySuite.IO;
-using Newtonsoft.Json;
+﻿using Vale.Geographic.Domain.Repositories.Interfaces;
+using Vale.Geographic.Domain.Base.Interfaces;
+using Vale.Geographic.Application.Services;
 using Vale.Geographic.Application.Base;
 using Vale.Geographic.Application.Dto;
-using Vale.Geographic.Application.Services;
-using Vale.Geographic.Domain.Base.Interfaces;
 using Vale.Geographic.Domain.Entities;
-using Vale.Geographic.Domain.Repositories.Interfaces;
 using Vale.Geographic.Domain.Services;
 using Vale.Geographic.Infra.Data.Base;
+using NetTopologySuite.Geometries;
+using System.Collections.Generic;
+using GeoJSON.Net.Geometry;
+using NetTopologySuite.IO;
+using GeoAPI.Geometries;
+using NetTopologySuite;
+using Newtonsoft.Json;
+using System.Linq;
+using AutoMapper;
+using System;
 
 namespace Vale.Geographic.Application.Core.Services
 {
     public class PointOfInterestAppService : AppService, IPointOfInterestAppService
     {
-        private readonly IPointOfInterestRepository _pointOfInterestRepository;
+        private readonly IPointOfInterestRepository pointOfInterestRepository;
 
         public PointOfInterestAppService(IUnitOfWork uoW, IMapper mapper, IPointOfInterestService pointOfInterestService,
             IPointOfInterestRepository pointOfInterestRepository) : base(uoW, mapper)
         {
             this.pointOfInterestService = pointOfInterestService;
-            this._pointOfInterestRepository = pointOfInterestRepository;
+            this.pointOfInterestRepository = pointOfInterestRepository;
         }
 
         public IPointOfInterestService pointOfInterestService { get; set; }
-
 
         public void Delete(Guid id)
         {
             try
             {
                 UoW.BeginTransaction();
-                var obj = pointOfInterestService.GetById(id);
+                PointOfInterest obj = pointOfInterestService.GetById(id);
+
+                if (obj == null)
+                    throw new ArgumentNullException();
+
                 pointOfInterestService.Delete(Mapper.Map<PointOfInterest>(obj));
                 UoW.Commit();
             }
@@ -48,7 +52,6 @@ namespace Vale.Geographic.Application.Core.Services
             }
         }
 
-
         public PointOfInterestDto GetById(Guid id)
         {
             return Mapper.Map<PointOfInterestDto>(pointOfInterestService.GetById(id));
@@ -56,21 +59,31 @@ namespace Vale.Geographic.Application.Core.Services
 
         public IEnumerable<PointOfInterestDto> GetAll(IFilterParameters parameters, out int total)
         {
-
-            var query = _pointOfInterestRepository
+            IEnumerable<PointOfInterest> query = pointOfInterestRepository
                 .GetAll(x => true, parameters,
-                    new string[] { "FirstName", "LastName", "Type" })
+                    new string[] { "CategoryId", "Status", "AreaId" })
                 .ApplyPagination(parameters, out total)
                 .ToList();
 
             return Mapper.Map<IEnumerable<PointOfInterestDto>>(query);
-
         }
 
-        public IEnumerable<PointOfInterestDto> Get(bool? active, Guid categoryId, IFilterParameters request, out int total)
+        public IEnumerable<PointOfInterestDto> Get(bool? active, Guid? Id, Guid? categoryId, Guid? areaId, double? latitude, double? longitude,  double? altitude, IFilterParameters request, out int total)
         {
-            return Mapper.Map<IEnumerable<PointOfInterestDto>>(_pointOfInterestRepository.Get(active, categoryId, request.sort,
-                request.page, request.per_page, out total));
+            IGeometry point = null;
+
+            if (longitude.HasValue && latitude.HasValue)
+            {
+                var coordenate = new GeoJSON.Net.Geometry.Point(new Position(latitude.Value, longitude.Value, altitude));
+                var json = JsonConvert.SerializeObject(coordenate);
+                var geometryFactory = NtsGeometryServices.Instance.CreateGeometryFactory(srid: 4326);
+
+                point = (Geometry)geometryFactory.CreateGeometry(new GeoJsonReader().Read<Geometry>(json));
+            }
+
+            IEnumerable<PointOfInterest> points = pointOfInterestRepository.Get( Id, out total, null, point, active, categoryId, areaId, request);
+
+            return Mapper.Map<IEnumerable<PointOfInterestDto>>(points);
         }
 
         public PointOfInterestDto Insert(PointOfInterestDto obj)
@@ -81,15 +94,48 @@ namespace Vale.Geographic.Application.Core.Services
 
                 PointOfInterest pointOfInterest = Mapper.Map<PointOfInterest>(obj);
 
-                var json = JsonConvert.SerializeObject(obj.Location);
+                var json = JsonConvert.SerializeObject(obj.Geojson.Geometry);
                 var geometryFactory = NtsGeometryServices.Instance.CreateGeometryFactory(srid: 4326);
-
                 pointOfInterest.Location = (Geometry) geometryFactory.CreateGeometry(new GeoJsonReader().Read<Geometry>(json));
+
                 pointOfInterest = pointOfInterestService.Insert(pointOfInterest);
 
                 UoW.Commit();
 
                 return Mapper.Map<PointOfInterestDto>(pointOfInterest);
+            }
+            catch (Exception ex)
+            {
+                UoW.Rollback();
+                throw ex;
+            }
+        }
+
+        public IEnumerable<PointOfInterestDto> Insert(CollectionPointOfInterestDto obj)
+        {
+            try
+            {
+                UoW.BeginTransaction();
+
+                List<PointOfInterest> points = new List<PointOfInterest>();
+
+                foreach (var feature in obj.Geojson.Features)
+                {
+                    PointOfInterest pointOfInterest = Mapper.Map<PointOfInterest>(feature);
+
+                    pointOfInterest.AreaId = obj.AreaId;
+                    pointOfInterest.CategoryId = obj.CategoryId;
+
+                    var json = JsonConvert.SerializeObject(feature.Geometry);
+                    var geometryFactory = NtsGeometryServices.Instance.CreateGeometryFactory(srid: 4326);
+                    pointOfInterest.Location = (Geometry)geometryFactory.CreateGeometry(new GeoJsonReader().Read<Geometry>(json));
+
+                    points.Add(pointOfInterestService.Insert(pointOfInterest));                       
+                }                              
+
+                UoW.Commit();
+
+                return Mapper.Map<IEnumerable<PointOfInterestDto>>(points);
             }
             catch (Exception ex)
             {
@@ -104,12 +150,24 @@ namespace Vale.Geographic.Application.Core.Services
             {
                 UoW.BeginTransaction();
 
-                var PointOfInterest = Mapper.Map<Domain.Entities.PointOfInterest>(obj);
-                PointOfInterest.Id = id;
-                PointOfInterest = pointOfInterestService.Update(PointOfInterest);
+                //var pointOfInterestOriginal = pointOfInterestService.GetById(id);
+                var pointOfInterestOriginal = pointOfInterestRepository.RecoverById(id);            
+                if (pointOfInterestOriginal == null)
+                    throw new ArgumentNullException();
+
+                PointOfInterest pointOfInterest = Mapper.Map<PointOfInterest>(obj);
+                pointOfInterest.Id = id;
+                pointOfInterest.CreatedAt = pointOfInterestOriginal.CreatedAt;
+
+                var json = JsonConvert.SerializeObject(obj.Geojson.Geometry);
+                var geometryFactory = NtsGeometryServices.Instance.CreateGeometryFactory(srid: 4326);
+                pointOfInterest.Location = (Geometry)geometryFactory.CreateGeometry(new GeoJsonReader().Read<Geometry>(json));
+
+                pointOfInterest = pointOfInterestService.Update(pointOfInterest);
 
                 UoW.Commit();
-                return Mapper.Map<PointOfInterestDto>(PointOfInterest);
+
+                return Mapper.Map<PointOfInterestDto>(pointOfInterest);
             }
             catch (Exception ex)
             {
